requirements.txt
langgraph>=0.2.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-google-genai>=2.0.0
langchain-community>=0.1.0
pydantic>=2.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
typing-extensions>=4.5.0
